requirements.txt
amqp==5.1.0
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.3
factory-boy==3.2.1
Faker==13.3.2
iniconfig==1.1.1
kombu==5.2.4
packaging==21.3
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytest-celery==0.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
